move warning indicator and equilibrium input to gui/util.py
